Moved collaborator view tests
